steering input needs to be an int, not a float

+++ find_cone.py
@@ -25,7 +25,7 @@ class App():
             # update error
             error = -fix.x
             print(f'error = {error} (degrees)')
-            await self.controller.set_steering(error * 32)
+            await self.controller.set_steering(int(error * 32))
         
     async def run(self):
         await self.controller.start()
